+++ fix/fixer.go
@@ -58,6 +58,7 @@ func init() {
 		"vsphere-iso-net-disk":       new(FixerVSphereNetworkDisk),
 		"iso-checksum-type-and-url":  new(FixerISOChecksumTypeAndURL),
 		"qemu-host-port":             new(FixerQEMUHostPort),
+		"azure-exclude_from_latest":  new(FixerAzureExcludeFromLatest),
 		"proxmox-type":               new(FixerProxmoxType),
 	}
 
@@ -94,6 +95,7 @@ func init() {
 		"vsphere-iso-net-disk",
 		"iso-checksum-type-and-url",
 		"qemu-host-port",
+		"azure-exclude_from_latest",
 		"proxmox-type",
 	}
 }